Remove deprecated InlineLink container, rename replacements
