--- src/app/containers/InlineLink/index.new.jsx
@@ -1,42 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-import pathToRegexp from 'path-to-regexp';
-import InlineLink from '../../components/InlineLink';
-import Blocks from '../Blocks';
-import fragment from '../Fragment';
-import { inlineLinkModelPropTypes } from '../../models/propTypes/inlineLink';
-import { articleRegexPath } from '../../routes';
-
-const InternalInlineLink = InlineLink.withComponent(Link);
-
-const componentsToRender = { fragment };
-
-const InlineLinkContainer = ({ locator, blocks }) => {
-  const regexp = pathToRegexp(articleRegexPath, [], {
-    start: false,
-    end: false,
-  });
-
-  const result = regexp.exec(locator);
-  // if URL matches a valid route, use a react-router link
-  if (result) {
-    // the path is the first item in the array
-    const path = result[0];
-    return (
-      <InternalInlineLink to={path}>
-        <Blocks blocks={blocks} componentsToRender={componentsToRender} />
-      </InternalInlineLink>
-    );
-  }
-
-  // else return a normal hyperlink
-  return (
-    <InlineLink href={locator}>
-      <Blocks blocks={blocks} componentsToRender={componentsToRender} />
-    </InlineLink>
-  );
-};
-
-InlineLinkContainer.propTypes = inlineLinkModelPropTypes;
-
-export default InlineLinkContainer;
